Unexpected token ۶ in JSON at position 0
Charlie,
Here is my update for the FAQ on this issue above that we talked about.  I will email you a document too with screen shots if you care to add that as well.  I did know how to attach a document into this document that I was editing in Github.

Thanks,
Ed Moore